fetch/api/request/request-consume.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Request consume</title>
    <meta name="help" href="https://fetch.spec.whatwg.org/#request">
    <meta name="help" href="https://fetch.spec.whatwg.org/#body-mixin">
    <meta name="author" title="Canon Research France" href="https://www.crf.canon.fr">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
    function checkBodyText(request, expectedBody) {
      return request.text().then( function(bodyAsText) {
        assert_equals(bodyAsText, expectedBody, "Retrieve and verify request's body");
        assert_true(request.bodyUsed === true, "body as text: bodyUsed turned true");
      });
    }

    function checkBodyBlob(request, expectedBody) {
      return request.blob().then(function(bodyAsBlob) {
        var promise = new Promise( function (resolve, reject) {
          var reader = new FileReader();
          reader.onload = function(evt) {
            resolve(reader.result)
          };
          reader.onerror = function () {
            reject("Blob's reader failed");
          };
          reader.readAsText(bodyAsBlob);
        });
        return promise.then(function(body) {
          assert_equals(body, expectedBody, "Retrieve and verify request's body");
          assert_true(request.bodyUsed === true, "body as blob: bodyUsed turned true");
        });
      });
    }

    function checkBodyArrayBuffer(request, expectedBody) {
      return request.arrayBuffer().then( function(bodyAsArrayBuffer) {
        var decoder = new TextDecoder("utf-8");
        var strBody = decoder.decode(bodyAsArrayBuffer);
        assert_equals(strBody, expectedBody, "Retrieve and verify request's body");
        assert_true(request.bodyUsed === true, "body as arrayBuffer: bodyUsed turned true");
      });
    }

    function checkBodyJson(request, expectedBody) {
      return request.json().then(function(bodyAsJson) {
        var strBody = JSON.stringify(bodyAsJson)
        assert_equals(strBody, expectedBody, "Retrieve and verify request's body");
        assert_true(request.bodyUsed === true, "body as json: bodyUsed turned true");
      });
    }

    function checkBodyFormData(request, expectedBody) {
      return request.formData().then(function(bodyAsFormData) {
        for (var name of expectedBody.keys())
          assert_equals(bodyAsFormData.get(name), expectedBody.get(name), "Retrieve and verify request's body" );
        assert_true(request.bodyUsed === true, "body as formData: bodyUsed turned true");
     });
    }

    function checkrequestBody(body, bodyType, checkFunction) {
      promise_test(function(test) {
        var request = new Request("", {"method": "POST", "body": body });
        assert_true(request.bodyUsed === false, "bodyUsed is false at init");
        return checkFunction(request, body);
      }, "Consume request's body as " + bodyType);
    }

    var formData = new FormData();
    formData.append("name", "value")
    checkrequestBody("This is request's body", "text", checkBodyText);
    checkrequestBody("This is request's body", "blob", checkBodyBlob);
    checkrequestBody("This is request's body", "arrayBuffer", checkBodyArrayBuffer);
    checkrequestBody(JSON.stringify("This is request's body"), "json", checkBodyJson);
    checkrequestBody(formData, "formData", checkBodyFormData);
    </script>
  </body>
</html>
